visual-viewport/viewport-segments.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Viewport Segments: visualViewport.segments</title>
<link rel="help" href="https://wicg.github.io/visual-viewport/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
"use strict";

promise_test(async t => {
  assert_true('segments' in visualViewport, "`segments` must be a property of visualViewport.");
  assert_equals(visualViewport.segments, null, "For a viewport not segmented, no segments must be exposed.");

  // HORIZONTAL
  await test_driver.set_viewport_orientation('horizontal');
  assert_equals(window.visualViewport.segments.length, 2);
  assert_true(matchMedia('(horizontal-viewport-segments: 1)').matches);
  assert_true(matchMedia('(vertical-viewport-segments: 2)').matches);
  assert_not_equals(visualViewport.segments, null, "For a viewport segmented, segments must be exposed.");
  console.log('JV666 visualViewport.segments: ' + JSON.stringify(visualViewport.segments, null, 2));

  // VERTICAL
  const segmentsMQL = window.matchMedia('(horizontal-viewport-segments: 2)');
  const promise = new Promise(resolve => {
    segmentsMQL.addEventListener(
      'change',
      () => { resolve(segmentsMQL.matches); },
      { once: true }
    );
  });
  await test_driver.set_viewport_orientation('vertical');
  assert_true(await promise)
  assert_true(matchMedia('(horizontal-viewport-segments: 2)').matches);
  assert_true(matchMedia('(vertical-viewport-segments: 1)').matches);
}, 'Viewport segments');
</script>
